Add linked nodes for dimension to python client
